Fixed custom wallpaper being not respected by sddm.
For some reason sddm cannot handle absolute file paths to wallpapers
and also needs the wallpaper to be readable by others.

This is fixed by copying the wallpaper to the root directory of the
selected theme.

On save the sddmauthhelper copies the background from the absolute path
into the theme directory and sets the "background" key of the
theme.user.conf to the copied file. If previously a different background
was
set it is removed beforehand.

REVIEW: 126524