Add SDDM theme install/uninstall with GHNS to the SDDM KCM
Summary:
Add SDDM theme install/uninstall with Get Hot New Stuff to the SDDM KCM.
Button arrangement is copied from the colour KCM. We can install from either a
local .zip file or the KDE store.

It assumes a zip file containing a folder/folders of themes to be installed
into the SDDM directory. metadata is very loosely checked.

As it needs to be root to install, this is added into the helper.

An additional executable is needed to make KNS work properly. We need to keep track of which
KNS packages contain which themes, which is done in a local config file.

Test Plan:
Installed a theme via GHNS
Uninstalled it
Installed a theme manually
Uninstalled it

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3685