Sync MAX_UID with upstream
Dynamic users are getting more widely used in latest versions of systemd, and some users are reporting them showing up in SDDM. These users get an UID in the 61184–65519 range.
Lower MAX_UID to 60000 to make sure they are not displayed in SDDM by default. This also brings this value in sync with the default upstream MaximumUID

Differential Revision: https://phabricator.kde.org/D13911