Crash in sddmthemeinstaller invalid use of errorString
Summary:
The destruction of KAuth::ExecuteJob* "job" is not handled properly when passed as a parameter to KMessageBox.

An intermediate QString is created containing the error text. This value is passed into the KMessageBox and qWarning() so that the object "job" is no longer an input. The console error is now first so an error message displayed even in the event of a GUI crash.

BUG: 406718
BUG: 404609
FIXED-IN: 5.16.0

Test Plan: The patch has been applied to my system and the crash no longer occurs.

Reviewers: ngraham, davidedmundson, #plasma

Reviewed By: ngraham, davidedmundson, #plasma

Subscribers: anthonyfieroni, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D20859