Check if sddm user exists
Summary:
Sddm-kcm should depend on sddm.
Sddm sets up an sddm user which is used to run the greeter.
Therefore it should not be possible to run sddm-kcm without an sddm user existing.

But if it should happen that's the case, this patch adds a check for the sddm user.
If the user is not present, everything will be aborted and an error message will pop-up.

openSUSE BUG:
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1145182
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1146817
  - https://bugzilla.opensuse.org/show_bug.cgi?id=1146831

Test Plan:
changed line `QString sddmHomeDirPath = KUser("sddm").homeDir();` to look up some bogus user
syncing and resetting is aborted, error message pops up

{F7273960}

Reviewers: #plasma, davidedmundson, ngraham, #vdg, fvogt

Reviewed By: ngraham, #vdg

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23378