Have authhelper define config file paths on its own
Summary:
Suggested for D26813. Prerequiste if user-manager is to not send any SDDM config file path on its own.

The config file paths are now defined in the authhelper as opposed to being sent in a variant map from the main kcm file.

I also edited paths in that main file because they were wrong.

This does not yet solve bug 414396.

CC-BUG: 342722

Test Plan: Testing with D26813 I managed to write to the config file without user-manager defining the path.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27142