Port to ManagedConfigModule
The settings are now read via KConfigXT, with a trick to follow the config
scheme of SDDM. One KConfig object is used to read the defaults and another one
to read the actual config values, see SddmSettingsBase.h for details.
The UI is now entirely in Qml and follows the GridviewKCM style. The preview pane
has been moved into a Dialog. The "Advanced" has been moved into a page called
"Behavior" and the sync options live now in a OverlaySheet accessible from the main
page as does the background configuration.
BUG:403530
BUG:407689
BUG:411004
BUG:411504
BUG:419327
BUG:424639