Use Kirigami.ActionToolBar to auto-resize the SDDM KCM's footer
Based on the work done for the Icons KCM, this change ensures that the
buttons at the bottom of the KCM actually fit in the window, by using
Kirigami.ActionToolBar's ability to collapse buttons which don't fit
into a popup menu style list of actions. Thanks to @ilyabizyae for
noticing this one.