Warn the user about KWallet complications when using auto-login
When auto-login is being used, The user's KWallet does not unlock
automatically if it has a password set, which can be annoying, and is a
technical detail that is not explained to the user.

This commit adds a warning message that tells all of this to the user,
gives them a suggested course of action (set the password to an empty
string), warns them of the consequences of this, and gives them a button
they can click that will take them to the place where they can do it.

The message is not shown when kwallet support is not installed.