Clear cache when syncing
The syncing process simply writes files to the SDDM user's homedir; it
doesn't do any of the fancy notifying that can happen when you change
settings in a running session. As a result, old cached stuff can be kept
around and leak into the user's settings.

Let's always clear the cache at the beginning of a sync to make sure
that we return to a fresh clean state.

BUG: 440957
FIXED-IN: 5.24.6


(cherry picked from commit 77052b081b709e3d7dd76f4c90c6deb952272950)