Avoid install prefix in config path lookup
CMAKE_INSTALL_FULL_SYSCONFDIR has unexpected behaviour when used with
prefixes.

>   read-only single-machine data (etc, or /etc if CMAKE_INSTALL_PREFIX
is /usr)

This is typically because prompting for root every install would be
annoying. Our use case here is not to install files but to refer to a
file in a known location once installed, so we do not want this
behaviour.

BUG: 440564
FIXED-IN: 5.26