Add optional manual input for autologin combobox
This allows the user to specify the login of an account that is outside
the uid range, and accordingly, is not in the drop-down list.

It also allows to display the login of an account that is already set as
an auto-login (for example manually in config file), which is outside the
range of the uid.