Fix needssave state management
When opening the "Advanced" page we first get onEnabledChanged with false and then with true
for the comboboxes.

This causes the KCM to be in modified state despite no changes being made.
When closing the KCM this further confuses systemsettings, causing a crash

To avoid this restructure the code so that settings are only written when interactively changing
the checkbox value

BUG: 490937

SENTRY: SYSTEMSETTINGS-2DD
(cherry picked from commit 5cb6a56f9084ee56fe64f3c3db7a424f63af3662)